fix: disable failing and fix stunted unit tests
