+++ src/shared/constants/index.ts
@@ -1,5 +1,9 @@
+/**
+ * @jest-environment jsdom
+ */
+
 import defaultAudioStillPath from '../../assets/images/audio-still.svg';
-import { tText } from '../helpers/translate';
+import { tText } from '../helpers/translate-text';
 
 export { LANGUAGES } from './languages';
 export { ROUTE_PARTS } from './routes';
