+++ src/shared/helpers/translate.ts
@@ -1,0 +1,6 @@
+/**
+ * @jest-environment jsdom
+ */
+
+export * from './translate-text';
+export * from './translate-html';
