+++ src/shared/translations/i18n.ts
@@ -1,9 +1,13 @@
+/**
+ * @jest-environment jsdom
+ */
+
 import I18n from 'i18next';
 import XHR from 'i18next-xhr-backend';
 import { lowerCase, upperFirst } from 'lodash-es';
 import { initReactI18next } from 'react-i18next';
 
-import { getEnv } from '../helpers';
+import { getEnv } from '../helpers/env';
 
 let resolveTranslations: (value?: unknown) => void | undefined;
 export const waitForTranslations = new Promise((resolve) => {
